test(ScreeningMovieActivityTest): AdViewHolder에 광고 이미지가 보이는지 테스트
